Make the seed job duty a responsibility of the RundeckLock object
